feat: added useWallet hook
