+++ packages/modal/src/account-linking/react.ts
@@ -3,17 +3,10 @@ export type {
   CitadelLinkAccountPayload,
   IUseLinkAccount,
   IUseSwitchAccount,
-  IUseWallets,
   LinkAccountParams,
   LinkAccountResult,
   LinkedAccountInfo,
   UnlinkAccountPayload,
   UnlinkAccountResult,
 } from "@web3auth/no-modal/account-linking/react";
-export {
-  makeAccountLinkingRequest,
-  makeAccountUnlinkingRequest,
-  useLinkAccount,
-  useSwitchAccount,
-  useWallets,
-} from "@web3auth/no-modal/account-linking/react";
+export { makeAccountLinkingRequest, makeAccountUnlinkingRequest, useLinkAccount, useSwitchAccount } from "@web3auth/no-modal/account-linking/react";
