+++ packages/modal/src/react/hooks/useWallets.ts
@@ -1,0 +1,2 @@
+export type { IUseWallets } from "@web3auth/no-modal/react";
+export { useWallets } from "@web3auth/no-modal/react";
