+++ packages/modal/src/vue/composables/useWallets.ts
@@ -1,0 +1,2 @@
+export type { IUseWallets } from "@web3auth/no-modal/vue";
+export { useWallets } from "@web3auth/no-modal/vue";
